寫了四張卡片
暫時想不到卡片堆疊邏輯
然後把角色和怪的血量和攻擊力改成float
不然沒辦法做乘法

+++ Assets/Monster/Flower/Flower.cs
@@ -4,10 +4,10 @@ public class Flower : MonoBehaviour
 {
     public float FlowerATKRange = 3.0f;
     public float FlowerATKCD = 1.0f;
-    public float FlowerSpeed = 1f;
-    public int FlowerHealth = 100;
+    public float FlowerSpeed = 1.0f;
+    public float FlowerHealth = 100.0f;
     float FlowerATKTimer = 0;
-    public void TakeFlowerDMG(int damage)
+    public void TakeFlowerDMG(float damage)
     {
         FlowerHealth -= damage;
     }
